requirements.txt
arabic-reshaper==3.0.0
asgiref==3.7.2
asn1crypto==1.5.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
contourpy==1.1.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.4
cssselect2==0.7.0
cycler==0.11.0
Django==4.2.5
django-cors-headers==4.2.0
django-extensions==3.2.3
djangorestframework==3.14.0
drf-extra-fields==3.7.0
filetype==1.2.0
fonttools==4.42.1
html5lib==1.1
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
kiwisolver==1.4.5
lxml==4.9.3
MarkupSafe==2.1.3
matplotlib==3.8.0
mysqlclient==2.2.0
numpy==1.26.0
oscrypto==1.3.0
packaging==23.1
Pillow==10.0.1
pycparser==2.21
pyHanko==0.20.1
pyhanko-certvalidator==0.24.1
pyparsing==3.1.1
pypdf==3.16.2
pypng==0.20220715.0
python-bidi==0.4.2
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
qrcode==7.4.2
reportlab==3.6.13
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
svglib==1.5.1
sweetify==2.3.1
tinycss2==1.2.1
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.0.1
uritemplate==4.1.1
uritools==4.0.2
urllib3==2.0.5
webencodings==0.5.1
xhtml2pdf==0.2.11
